Changed autoprobe behavior to average instead of exact match
This version probes every site 5 times and reports the average of the 5 as the measurement rather than waiting for an exact match across 21 probings. It is possible that probing may take a small bit longer, but for many who's probe are not repeatable enough to match in fewer than 5 probings this may be the only way to complete the process and get useable (not perfect) results.