Changed probe average method to a mode/median filter
The mode/median filter first tests for an exact match, as the original autocalibration method did. If it finds an exact match to the first probe, it will run as normal, i.e. cut short the probing and move to the next spot. If it does not find an exact match to the first probe, it will construct an array of the probe points (set to do 7 probes) and test for the mode value. If it finds a mode it will report that, if no mode is found or if a bimodal situation is found, it will report the median instead. There are some diagnostic reports to tell what the individual probe points are, as well as if it's an exact match or mode/median and what is the range of the probes for each site.