Buffer Overflow in Z_probe()
The saved_position variable is an array with a length of 3. NUM_AXIS
typically include an extruder on top of the 3 axis.
This cannot end well… This bug doesn’t seem to affect everybody but I
was experiencing some nasty effect with my configuration (I am using a
Mac and a Arduino Mega) where after Z_probing the Y axis was shorten
and the head crashed on the bed. The fix is not pretty but follow the
previous block of code.